Fixed USPTO Inventors

+++ routes/client/company.js
@@ -1334,17 +1334,20 @@ route.post("/", [authJWT.verifyToken, clientDBConnection.connect], async(req, re
  */
 route.delete("/", [authJWT.verifyToken, clientDBConnection.connect], async(req, res, next) => {
     try{
-        let IDs = req.query.companies;
-        if(IDs.length > 0) {
-            IDs = JSON.parse(IDs)
+        let {companies, type}  = req.query;
+        
+        if(companies.length > 0) {
+            companies = JSON.parse(companies)
             const Representative = req.connection_db.define('Representatives', Representatives.mainStructure, Representatives.options);
+            
             const findCompanies = await Representative.findAll({
-                attributes:['representative_id', 'parent_id', 'original_name'],
-                where:{representative_id: IDs},
+                attributes:['representative_id', 'parent_id', 'original_name', 'type'],
+                where:{representative_id: companies},
                 group:['representative_id','parent_id']
             });
             const updateKPICompanies=[],  deleteParentCompanies = [], reUpdateCompanies = [], deleteCompanies = [], activityLogs = [], currentDate = moment(new Date()).format('YYYY-MM-DD hh:mm:ss');
             if(findCompanies.length > 0) {
+                
                 const promise = findCompanies.map(c => {
                     if(c.parent_id == 0) {
                         deleteParentCompanies.push(c.representative_id);
@@ -1352,7 +1355,7 @@ route.delete("/", [authJWT.verifyToken, clientDBConnection.connect], async(req, 
                     } else {
                         if(!updateKPICompanies.includes(c.parent_id)){
                             updateKPICompanies.push(c.parent_id); 
-                            reUpdateCompanies(c.parent_id);
+                            reUpdateCompanies.push(c.parent_id);
                         }
                     }
                     deleteCompanies.push(c.representative_id);
@@ -1366,6 +1369,8 @@ route.delete("/", [authJWT.verifyToken, clientDBConnection.connect], async(req, 
                 });
 
                 await Promise.all(promise);
+
+                
 
                 if(deleteParentCompanies.length > 0) {
                     const findParentSubCompanies = await Representative.findAll({
@@ -1384,10 +1389,21 @@ route.delete("/", [authJWT.verifyToken, clientDBConnection.connect], async(req, 
 
 
                 if(deleteCompanies.length > 0) {
+                    let destroyAllCompanies = null
+                    if(typeof type != 'undefined' && type == 1) {
+                        /**
+                         * Keep all the companies outside group and delete group
+                         */
+                        destroyAllCompanies = await  Representative.update({parent_id: 0},{
+                            where: {representative_id: deleteCompanies},
+                        })
+                    } else {
+                        destroyAllCompanies = await  Representative.destroy({
+                            where: {representative_id: deleteCompanies},
+                        })
+                    }
                     
-                    const destroyAllCompanies = await  Representative.destroy({
-                        where: {representative_id: deleteCompanies},
-                    })
+                    
 
                     if(destroyAllCompanies != null) {
                         ActivityLogs.bulkCreate(activityLogs);
@@ -1396,32 +1412,6 @@ route.delete("/", [authJWT.verifyToken, clientDBConnection.connect], async(req, 
                                 where: {representative_id: deleteParentCompanies, organisation_id: req.orgId},
                             });
                             console.log("destroyAllTransactions", destroyAllTransactions);
-                            if(destroyAllTransactions) {
-                                /**
-                                 * Delete KPI counter, Tree, Timeline, Error
-                                 */
-                                //remove from list 1, list 2, assets, transactions
-
-                                /* await Validity.destroy({
-                                    where: {representative_id: deleteParentCompanies, organisation_id: req.orgId},
-                                });
-                                await Transactions.destroy({
-                                    where: {representative_id: deleteParentCompanies, organisation_id: req.orgId},
-                                });
-                                await TreeParties.destroy({
-                                    where: {representative_id: deleteParentCompanies, organisation_id: req.orgId},
-                                });
-                                await TreePartiesCollections.destroy({
-                                    where: {representative_id: deleteParentCompanies, organisation_id: req.orgId},
-                                });
-                                await Errors.destroy({
-                                    where: {representative_id: deleteParentCompanies, organisation_id: req.orgId},
-                                });
-
-                                await Timelines.destroy({
-                                    where: {representative_id: deleteParentCompanies, organisation_id: req.orgId},
-                                }); */
-                            }
                         }
